Move interfaces to types folder
