+++ components/documents/FolderFilesModal.tsx
@@ -8,37 +8,12 @@ import {
   Loader2,
 } from "lucide-react";
 import { decodeFileName } from "@/utils/decodeFileName";
-
-export interface DocumentItem {
-  id: string;
-  name: string;
-  category: string;
-  type: string;
-  size: string;
-  updated: string;
-  url: string;
-}
-
-export interface CategoryCount {
-  name: string;
-  count: number;
-}
-
-export interface StorageStats {
-  total: string;
-  pdf: string;
-  docs: string;
-  sheets: string;
-}
-
-interface DocumentManagementPageProps {
-  documents: DocumentItem[];
-  recentDocuments: DocumentItem[];
-  categories: CategoryCount[];
-  storage: StorageStats;
-  onDelete: (id: string) => void;
-  onUpload: (files: FileList) => Promise<void>;
-}
+import type {
+  DocumentItem,
+  CategoryCount,
+  StorageStats,
+  DocumentManagementPageProps,
+} from "@/types/documents";
 
 export default function DocumentManagementPage({
   documents,
