+++ types/mock.ts
@@ -1,0 +1,4 @@
+export interface TestEntity {
+  id: string;
+  name: string;
+}
